refactor: move method to more appropriate location
The hcloud.Setenv() method is needed in multiple packages and better
belongs into internal/testsupport.
Keeping it in hcloud causes cyclical imports for soon to follow tests.

hcloud.SkipEnv() is unused and removed.
